==> 1_setup/stressors/1_process_shipping_data.Rmd <==
---
title: "Process shipping stressors"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(sf)
library(oharac) ### remotes::install_github('oharac/oharac')
  ### includes some helper functions, e.g., dt_join
library(tidyverse)
library(here)

source(here('common_fxns.R'))
source('stressor_fxns.R') ### in same dir as this Rmd

```

# Summary

Process shipping stressors, including a layer for ship strikes and for noise pollution, to 10 km Mollweide CRS.  Also process benthic structures, including oil, gas, and wind farms. 

* Commercial/passenger ship combined layer for large ships that contribute to whale strikes and (potentially) chemical pollution
* Fishing/leisure craft combined layer for smaller ships (stressors?)
* Combined layer of all ship tracks (stressors?)

Note: Ship density is NOT log-transformed, in a departure from prior CHI methods - e.g., Halpern et al. 2019.  Since passage of a ship does not likely leave persistent impacts (e.g., once a ship has passed, it can't still inflict wildlife strikes), leaving it in its raw density seems appropriate.  

We may wish to revisit this with respect to ship-based chemical pollutants etc.

# Data

* Data from: https://datacatalog.worldbank.org/search/dataset/0037580
    * Cite as: Cerdeiro, D. A., Komaromi, A., Liu, Y., & Saeed, M. (2020). World Seaborne Trade in Real Time: A Proof of Concept for Building AIS-based Nowcasts from Scratch (IMF Working Paper WP/20/57). International Monetary Fund.
* This is a count of ships per cell - aggregate, convert to ships per ocean area, then reproject.


# Methods

## Examine shipping density data

Files downloaded from https://datacatalog.worldbank.org/search/dataset/0037580 and unzipped to /home/shares/ohi/spp_vuln/spp_vuln_mapping/ship_density.  The data are in GeoTIFF raster format; each raster is about 10 GB.  Also included are .tif.ovr files, which are pyramids; these are 3 GB but can probably be safely discarded.

```{r}
ship_raw_dir <- here_anx('/stressors/ship_density/raw')
raw_tifs <- list.files(ship_raw_dir, pattern = '.tif$', full.names = TRUE)
comm_shipping_r <- raster(raw_tifs[1])
# class      : RasterLayer 
# dimensions : 33998, 72006, 2448059988  (nrow, ncol, ncell)
# resolution : 0.005, 0.005  (x, y)
# extent     : -180.0153, 180.0147, -84.98735, 85.00265  (xmin, xmax, ymin, ymax)
# crs        : +proj=longlat +datum=WGS84 +no_defs 
# source     : ShipDensity_Commercial1.tif 
# names      : ShipDensity_Commercial1 
# values     : 0, 62652999  (min, max)

### Check value to compare reprojected commercial layer - 
### sum of global ship tracks
# sum(values(comm_shipping_r), na.rm = TRUE)
# [1] 6.520743e+14
```

## Mask shipping rasters to ocean only

Because land cells are filled with zero, mask with ocean to convert these to NA.  This will ensure that aggregation up to 0.1° cells (~ 10 km at equator for better projection to 10 km Mollweide) will not include land in averaging.

Loop over each raw layer, save out masked layer.
```{r}
ship_masked_dir <- here_anx('stressors/ship_density/masked')

new_fs <- basename(raw_tifs) %>%
  tolower() %>%
  str_remove_all('[0-9]') %>%
  str_replace('.tif', '_masked.tif')
tifs_masked <- file.path(ship_masked_dir, new_fs)

### if files to be processed, create ocean mask
if(any(!file.exists(tifs_masked))) {
  ocean_mask_sf <- read_sf(here('_spatial/ne_10m_ocean/ne_10m_ocean.shp'))
  ocean_mask_r <- fasterize::fasterize(ocean_mask_sf, comm_shipping_r)

  # plot(ocean_mask_r, axes = FALSE, legend = FALSE)
}


for(i in seq_along(raw_tifs)) {
  ### i <- 1
  tif_raw <- raw_tifs[i]
  tif_masked <- tifs_masked[i]
  if(!file.exists(tif_masked)) {
    message('Masking ', basename(tif_raw))
    r <- raster(tif_raw)
    r_masked <- mask(r, ocean_mask_r, progress = 'text')
    message('Writing ', basename(tif_masked))
    writeRaster(r_masked, tif_masked, 
                progress = 'text',
                overwrite = TRUE)
  } else {
    message('File ', basename(tif_masked), ' exists... skipping!')
  }
}
```

## Aggregate to 0.1° in prep for projection

Convert shipping to tracks per km^2^ then aggregate.  Loop over each masked layer, save out aggregated layer.

```{r}
ship_agg_dir <- here_anx('stressors/ship_density/aggregated')

new_fs <- basename(tifs_masked) %>%
    str_replace('.tif', '_agg.tif')
tifs_agg <- file.path(ship_agg_dir, new_fs)

for(i in seq_along(tifs_masked)) {
  ### i <- 1
  tif_masked <- tifs_masked[i]
  tif_agg <- tifs_agg[i]
  
  if(!file.exists(tif_agg)) {
    message('Converting to ships per km^2: ', basename(tif_masked))
    r <- raster(tif_masked)
    
    ### convert to ships per km^2
    if(!exists('area_r')) {
      ### area in km^2
      area_r <- raster::area(r)
    }
    r_intensity <- r / area_r
    
    ### aggregate
    message('Aggregating to 0.1°: ', basename(tif_masked))
    r_agg <- aggregate(r_intensity, fact = 20, fun = mean, progress = 'text')
    
    message('Writing ', basename(tif_agg))
    writeRaster(r_agg, tif_agg, 
                progress = 'text',
                overwrite = TRUE)
  } else {
    message('File ', basename(tif_agg), ' exists... skipping!')
  }
}
for(t in tifs_agg) {
  ### t <- tifs_agg[1]
  r_name <- basename(t) %>% str_remove_all('shipdensity_|_masked_agg.tif')
  r <- raster(t)
  plot(r, axes = F, main = sprintf('%s tracks per km^2', r_name))
}
```

## Reproject to 10km x 10km Mollweide

Aggregated files are here, with values of ships per km^2^:

* dir: `/home/shares/ohi/spp_vuln/spp_vuln_mapping/stressors/ship_density/aggregated`
    * files: 
        * `shipdensity_commercial_masked_agg.tif`
        * `shipdensity_fishing_masked_agg.tif`
        * `shipdensity_leisure_masked_agg.tif`
        * `shipdensity_oilgas_masked_agg.tif`
        * `shipdensity_passenger_masked_agg.tif`
        
Create several layers: 

* a commercial and passenger ship layer, for ship strikes
    * ship strikes are primarily caused by huge ships, while smaller craft can avoid mammals more readily.
    * what about FPSOs in oil and gas layer? Include oil and gas layer as well.  It includes stationary rigs, but those should be a small contribution.
* a fishing and leisure craft layer, in case we want to separately analyze impacts of small(er) craft (even the largest fishing boats and yachts are barely half the size of an average cruise or container ship)
* an unweighted total tracks layer, just in case.
    
```{r}
### Load ocean proportion in Mollweide 10 km x 10 km (analysis CRS and res)
ocean_rast <- raster(here('_spatial/ocean_area_mol.tif'))
```

```{r}
dir_shipping <- here_anx('stressors/ship_density/aggregated')
ship_raw_fs <- list.files(dir_shipping, pattern = '.tif$', full.names = TRUE)

ship_raw_fs_large <- ship_raw_fs[str_detect(basename(ship_raw_fs), 'commercial|oilgas|passenger')]
ship_raw_fs_small <- ship_raw_fs[!ship_raw_fs %in% ship_raw_fs_large]
```

Define a function to project a stack of shipping density rasters (provided at 0.1 degree resolution) to Mollweide; convert to shipping track count; rescale.

```{r}
process_shipping_stack <- function(s) {
    ### Already masked and aggregated in prep script; reproject
  ship_stack <- projectRaster(s, ocean_rast, progress = 'text')
  
  ### projectRaster with ngb introduces some weird negative values; drop'em
  values(ship_stack)[values(ship_stack) < 0] <- 0
  ### convert densities to ship track count
  ship_stack_n <- ship_stack * 100 * ocean_rast
  
  ### check sum of commercial stack to make sure process is clean
  # sum(values(ship_stack_n[[1]]), na.rm = TRUE)
  # 6.453032e+14 # very close to original: 6.520743e+14

  ### summarize ship track counts and rescale
  ship_stack_r <- calc(ship_stack_n, fun = sum)
  ship_stack_rescaled_r <- rescale_stressor(ship_stack_r, qtile = .999)
  
  return(ship_stack_rescaled_r)
}
```

```{r create large ship layer}
ship_large_f <- here('_data/stressors_mol/shipping_large_2021.tif')
### unlink(ship_large_f)

if(!file.exists(ship_large_f)) {

  ship_l_stack <- stack(ship_raw_fs_large)
  # class      : RasterStack 
  # dimensions : 1700, 3601, 6121700, 3  (nrow, ncol, ncell, nlayers)
  # resolution : 0.1, 0.1  (x, y)
  # extent     : -180.0153, 180.0847, -84.99735, 85.00265  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # names      : shipdensity_commercial_masked_agg, shipdensity_oilgas_masked_agg, shipdensity_passenger_masked_agg 
  # min values :                                 0,                             0,                                0 
  # max values :                        53158596.0,                      863617.5,                           7743.0   
  
  ### summarize ship track counts and rescale
  ship_large_r_n <- process_shipping_stack(ship_l_stack)

  ### round to drop ludicrous precision then save out
  values(ship_large_r_n) <- round(values(ship_large_r_n), 5)
  writeRaster(ship_large_r_n, ship_large_f, overwrite = TRUE)
}

r <- raster(ship_large_f)
  
plot(r, axes = FALSE,
     main = 'Commercial, passenger, and oil/gas shipping, rescaled')

hist(r[r > 0], main = 'distribution of non-zero values')
```

```{r create small ship layer}
ship_small_f <- here('_data/stressors_mol/shipping_small_2021.tif')
### unlink(ship_small_f)

if(!file.exists(ship_small_f)) {

  ship_sm_stack <- stack(ship_raw_fs_small)
  # class      : RasterStack 
  # dimensions : 1700, 3601, 6121700, 2  (nrow, ncol, ncell, nlayers)
  # resolution : 0.1, 0.1  (x, y)
  # extent     : -180.0153, 180.0847, -84.99735, 85.00265  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # names      : shipdensity_fishing_masked_agg, shipdensity_leisure_masked_agg 
  # min values :                              0,                              0 
  # max values :                       847594.4,                      2801699.0 

  ### summarize ship track counts and rescale
  ship_small_r_n <- process_shipping_stack(ship_sm_stack)

  ### round to drop ludicrous precision then save out
  values(ship_small_r_n) <- round(values(ship_small_r_n), 5)
  writeRaster(ship_small_r_n, ship_small_f, overwrite = TRUE)
}

r <- raster(ship_small_f)
  
plot(r, axes = FALSE,
     main = 'Fishing and leisure vessels, rescaled')

hist(r[r > 0], main = 'distribution of non-zero values')
```

```{r create combined unweighted ship layer}
ship_all_unweighted_f <- here('_data/stressors_mol/shipping_all_unweighted_2021.tif')
### unlink(ship_all_unweighted_f)

if(!file.exists(ship_all_unweighted_f)) {

  ship_all_stack <- stack(ship_raw_fs)

  ### summarize ship track counts and rescale
  ship_all_r_n <- process_shipping_stack(ship_all_stack)

  ### round to drop ludicrous precision then save out
  values(ship_all_r_n) <- round(values(ship_all_r_n), 5)
  writeRaster(ship_all_r_n, ship_all_unweighted_f, overwrite = TRUE)
}

r <- raster(ship_all_unweighted_f)

plot(r, axes = FALSE,
     main = 'All vessels, unweighted, rescaled')

hist(r[r > 0], main = 'distribution of non-zero values')
```
