==> 1_setup/stressors/2_process_shipping_benthic_structures.Rmd <==
---
title: "Process benthic structure stressors"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(sf)
library(oharac) ### remotes::install_github('oharac/oharac')
  ### includes some helper functions, e.g., dt_join
library(tidyverse)
library(here)

source(here('common_fxns.R'))
source('stressor_fxns.R') ### in same dir as this Rmd

```

# Summary

Process benthic structures, including oil, gas, and wind farms.  These may incorporate data from World Bank ship tracks, persistent oceanic lights, and wind farm maps.  Estimate the footprint of structures and use this to determine relative impact on benthic habitats.

# Data

* Oil and gas structures: Data from: https://datacatalog.worldbank.org/search/dataset/0037580 (oil and gas layer)
    * Cite as: Cerdeiro, D. A., Komaromi, A., Liu, Y., & Saeed, M. (2020). World Seaborne Trade in Real Time: A Proof of Concept for Building AIS-based Nowcasts from Scratch (IMF Working Paper WP/20/57). International Monetary Fund.
    * supplement with light data?
* Offshore wind farms: Data from: https://figshare.com/articles/dataset/Global_offshore_wind_farm_dataset/13280252/5
    * Cite as: Zhang, T., Tian, B., Sengupta, D., Zhang, L., & Si, Y. (2021). Global offshore wind turbine dataset. Scientific Data, 8(1), 191. https://doi.org/10.1038/s41597-021-00982-z

# Methods

## Oil rigs: Identify potential vessel types

From the .csvs of data over time, pull the 2020 data files, filter to just vessel types noting "rig" and the like.

``` {r}
vessel_data_dir <- "/home/shares/ohi/stressors_2021/_raw_data/Cerdeiro_2020_vessel_density/grid_message_count_annual/"

vessel_2020_fs <- list.files(file.path(vessel_data_dir, 'grid_message_count_2020'), full.names = TRUE)

vessel_type_df <- data.frame(f = vessel_2020_fs) %>%
  mutate(type = str_remove_all(basename(f), 'grid_message_count_2020_|.csv'),
         type = tolower(type))

### types: platform, rig, jack_up; consider mooring_buoy
rigs <- vessel_type_df %>%
  filter(str_detect(type, 'platform$|rig$|jack_up$|mooring_buoy$'))
```

### Prelim examination of data

Read in the data for apparent rigs.  Convert the resulting points to a raster, and check for apparent ship tracks. 
Note we can aggregate here to 0.1 degree, approximately the resolution of our Mollweide 10 km analysis resolution, to greatly reduce processing intensity.

``` {r}
rig_data <- lapply(rigs$f, FUN = function(f) {
    x <- read_csv(f, show_col_types = FALSE) %>%
      janitor::clean_names() %>%
      mutate(across(-vessel_type, as.numeric))
    }) %>%
  bind_rows()

```

``` {r}
xyz <- rig_data %>%
  rename(x = round_lon, y = round_lat) %>%
  mutate(x = round(x, 1), y = round(y, 1)) %>%
  group_by(x, y) %>%
  summarize(z = sum(count_messages))

r_big <- rasterFromXYZ(xyz, res = 1, crs = '+init=epsg:4326', digits = 0)

plot(r_big, col = hcl.colors(20),
     main = '1.0° cells with observations (for easier visibility)',
     legend = FALSE, axes = FALSE)

r <- rasterFromXYZ(xyz, res = 0.1, crs = '+init=epsg:4326', digits = 1)

plot(r, col = hcl.colors(20),
     main = '0.1 deg cells with observations',
     legend = FALSE, axes = FALSE)
```

### Clean the data further

Drop observations where `count_messages` <= 5 (arbitrary but appears to work) to reduce or eliminate moving rigs passing through a cell.  For each remaining raw site (at 0.005 resolution), round the resolution, sum up how many raw sites are contained in the aggregated cell.  Use this to estimate the density of stationary rigs in a particular cell.

``` {r}
xyz2 <- rig_data %>%
  rename(x = round_lon, y = round_lat) %>%
  filter(count_messages > 5) %>%
  mutate(x = round(x, 1), y = round(y, 1)) %>%
  group_by(x, y) %>%
  summarize(z = n())

r_big <- rasterFromXYZ(xyz2, res = 1, crs = '+init=epsg:4326', digits = 0)

plot(r_big, col = hcl.colors(20), 
     main = '1.0° cells with observations (for easier visibility)',
     legend = FALSE, axes = FALSE)

r2 <- rasterFromXYZ(xyz2, res = 0.1, crs = '+init=epsg:4326', digits = 1)

plot(r2, col = hcl.colors(20), 
     main = '0.1 deg cells with observations',
     legend = FALSE, axes = FALSE)
```

### Transform points to Mollweide

From the overall rigs dataframe, filter out locations with fewer than five pings, then what is left is presumed to be stationary (multiple pings from the same site). Convert to a simple features points object, transform to Mollweide, and save out.

Values in result are:

`round_lat | round_lon | vessel_type`

```{r}

rigs_gpkg <- here_anx('stressors/benthic_structures/rigs_2020_mol.gpkg')
xyz3 <- rig_data %>%
  filter(count_messages > 5) %>%
  select(-count_messages) %>%
  distinct()

rigs_sf <- st_as_sf(xyz3, coords = c('round_lon', 'round_lat'),
                    crs = '+init=epsg:4326') %>%
  st_transform(crs = '+proj=moll')

write_sf(rigs_sf, rigs_gpkg)
```

Note that this could also be multiple ship tracks that happen to intersect, but based on type and lack of apparent motion, we arrive at `r nrow(xyz3)` rigs, which is pretty close to the 1470 figure noted here: https://www.weforum.org/agenda/2015/10/where-are-the-worlds-oil-rigs/.

## Offshore wind turbines

Read in offshore wind turbine data from https://figshare.com/articles/dataset/Global_offshore_wind_farm_dataset/13280252/5.  The shapefile is in:

* `stressors/benthic_structures/global_offshore_wind_turbines_dataset_v1.3/GOWF_V1.3.shp`

Read in and transform to Mollweide, then save out.

```{r}
owt_gpkg <- here_anx('stressors/benthic_structures/owt_2020_mol.gpkg')
owt_f <- here_anx('stressors/benthic_structures', 
                  'global_offshore_wind_turbines_dataset_v1.3',
                  'GOWF_V1.3.shp')
owt_sf <- read_sf(owt_f)
owt_mol_sf <- owt_sf %>%
  st_transform(crs = '+proj=moll')
write_sf(owt_mol_sf, owt_gpkg)
```

## Rasterize rig and turbine points using count

Then sum the two structure layers, weighting oil rigs and wind turbines approximately according to their respective footprints.  Assume an oil rig footprint is on the order of 1 ha (100 m x 100 m).  A single wind turbine is far smaller.  Both structures are likely to use scour protection around the base, assume ~ 25 m diameter.  

* Model oil rig as 100x100m with 25m buffer for scour protection, or about 150m x 150x = 22500 m^2^.
* model wind turbine as 10m diameter (for a big one!) with 25m buffer for scour protection, or $\pi$*30m^2^ = 2827 m^2^.

Then divide the impacted area by the ocean area of the cell, log transform (assume marginal impact of additional structures is decreasing) for a final stressor value.

```{r}
ocean_rast_mol <- raster(here('_spatial/ocean_area_mol.tif'))

rig_footprint <- 150 * 150
rigs_sf <- read_sf(rigs_gpkg) %>%
  select(geom)
rigs_r <- raster::rasterize(rigs_sf, ocean_rast_mol, fun = 'count', background = 0) * rig_footprint

owt_footprint <- 30^2 * pi
owt_sf <- read_sf(owt_gpkg) %>%
  select(geom)
owt_r <- raster::rasterize(owt_sf, ocean_rast_mol, fun = 'count', background = 0) * owt_footprint

### calculate total footprint of structures in cell, in m^2
benth_footprint_r <- rigs_r + owt_r
benth_footprint_r <- benth_footprint_r %>%
  mask(ocean_rast_mol)

writeRaster(benth_footprint_r, here_anx('stressors/benthic_structures',
                                        'structure_footprint_m2.tif'),
            overwrite = TRUE)

### Calculate proportional footprint, then rescale and save as stressor
benth_str <- (benth_footprint_r / (1e8 * ocean_rast_mol)) %>%
  log_plus() %>%
  rescale_stressor()

benth_f <- here('_data/stressors_mol/benth_str_2020.tif')
writeRaster(benth_str, benth_f, overwrite = TRUE)

```


``` {r}
r <- raster(benth_f)
  
plot(r, axes = FALSE,
     main = 'Benthic structures, rescaled')

hist(r[r > 0], main = 'distribution of non-zero values')
```
