==> 1_setup/stressors/3_process_fishing_bycatch.Rmd <==
---
title: "Watson Data: discards/bycatch"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(sf)
library(tidyverse)
library(here)
library(RColorBrewer)
library(xlsx)

library(oharac)
source(here('common_fxns.R'))
source('stressor_fxns.R') ### in same dir as this Rmd
```

# Summary

Read in data from Watson (2018).

# Data

Data from

* Watson, R.A. and Tidd, A.N. (2018) Mapping nearly a century and a half of global marine fishing: 1869 to 2015. Marine Policy 93, 171-177
* Watson, R. (2017) A database of global marine commercial, small-scale, illegal and unreported fisheries catch 1950-2014. Nature Scientific Data 4 (170039).

Included in data:

* Industrial Catch (1950 - 2017) - reported, iuu, and discard catch data for each cell location and unique identifier
* Non-Industrial Catch (1950 - 2017) - reported, iuu, and discard catch data for each cell location and unique identifier
* DATA CODE DEFINITIONS (gear/taxa/country codes and cell lat/lon references)

## How to download the data code definitions. 

For this, it is saved as an .xlsx file (Codes.xlsx). This file has 4 sheets in it which contain meta data that explain columns in the Index files, and one sheet with is "Spatial Cells Reference - contains geospatial information associated wtih the Industrial Catch data". 
To download this data, go to the IMAS website: http://data.imas.utas.edu.au/portal/search?uuid=ff1274e1-c0ab-411b-a8a2-5a12eb27f2c0 and dowload it manually.

This .xlsx file contains these sheets:

* Spatial Cells Reference (Cells) - contains geospatial information associated wtih the Industrial Catch data
* Gear Reference (Gear) - contains information regarding how different fishing gear is classified in the index datasets.
* Taxa Reference (Taxa) - contains information regarding how different taxa is classified in the index datasets.
* Country Reference (Country) - contains informations regarding how different countries are labeled in the index datasets. 

# Methods

## Access Watson data stored for OHI 2021

For our purposes we will only be using the most recent set of data.  These are stored on Mazu courtesy of Gage for OHI 2021.

* dir: `git-annex/globalprep/_raw_data/IMAS_GlobalFisheriesLandings/d2020`
    * file: `Catch2015_2019.rds`

```{r}
watson_dir <- here_anx('../../git-annex/globalprep/_raw_data', 
                       'IMAS_GlobalFisheriesLandings/d2020')
watson_df <- readRDS(file.path(watson_dir, 'Catch2015_2019.rds')) %>%
  janitor::clean_names()

codes_xlsx <- file.path(watson_dir, 'Codes_raw.xlsx')
# codes_sheets <- readxl::excel_sheets(codes_xlsx)
cell_df <- readxl::read_excel(codes_xlsx, sheet = 'Cell') %>%
  janitor::clean_names() %>%
  select(-c(x5:x7))
taxa_df <- readxl::read_excel(codes_xlsx, sheet = 'Taxa') %>%
  janitor::clean_names() %>%
  select(-c(x8:taxonkey_11), taxonkey = taxonkey_1)
ctry_df <- readxl::read_excel(codes_xlsx, sheet = 'Country') %>%
  janitor::clean_names() %>%
  select(cnumber = cnumber_1, fao_name)
gear_df <- readxl::read_excel(codes_xlsx, sheet = 'Gear') %>%
  janitor::clean_names() %>%
  select(-x9, -x10)

```

## Process bycatch from discards

Watson data includes discard estimates - "Discard catch rates that were estimated were not expected to be the same taxon as the reported landed taxon for each record, but rather would be comprised of a variety of taxa, targeted and non-targeted, many of which would not be represented in the original landings source data."  Through this, discards in any cell can be summed up across all targeted species, countries, and gear type.

However, we wish to include estimates for bycatch of demersal/pelagic species separately from pelagic, since we can account for water column position of all the species potentially impacted.  To do this, we can examine total bycatch across all targeted species and country, but separate gear type into benthic/demersal and pelagic/midwater types.

### Gears with high bycatch - from data

From the data, examine which gear types have a high ratio of discards to catch, across all cells as a rough first pass.  Drop the duplicate gear codes due to FAO coding and naming system...

```{r, results = 'asis'}
gear_clean_df <- gear_df %>%
  select('gear', 'f_gear_code', 'f_gear_label', 'fleet_gear_name') %>%
  distinct() %>%
  mutate(f_gear_label = tolower(f_gear_label),
         fleet_gear_name = tolower(fleet_gear_name),
         fleet_gear_name = str_replace(fleet_gear_name, 'tuna', ' tuna'),
         fleet_gear_name = str_replace_all(fleet_gear_name, '[^a-z]+', '_'))
catch_gear_df <- watson_df %>%
  oharac::dt_join(gear_clean_df, by = c('gear', 'f_gear_code'), type = 'left')

discard_catch_ratio <- catch_gear_df %>%
  group_by(fleet_gear_name) %>%
  summarize(discards  = sum(discards_ind, na.rm = TRUE),
            catch     = sum(reported_ind, na.rm = TRUE),
            d_c_ratio = round(discards / catch, 3),
            .groups = 'drop') %>%
  mutate(discards = round(discards),
         catch = round(catch)) %>%
  arrange(desc(d_c_ratio))

knitr::kable(discard_catch_ratio)

benthic_gears <- c('trawl', 'dredge', 'trap')

gear_d_c_ratio <- discard_catch_ratio %>%
  mutate(water_col = ifelse(fleet_gear_name %in% benthic_gears, 'benth', 'pel'))

write_csv(gear_d_c_ratio, here('int/watson_discard_catch_ratio.csv'))
```

We do not necessarily need to divide into high bycatch and low bycatch gears, since we have the actual discards reported.  If we can get effort information (via GFW) for all these gears, then we can adjust bycatch by CPUE and gear selectivity as a proxy for biomass of bycaught species.  If not, focus on those we can identify with effort, and those with the highest impacts.

The data are already presented in 0.5° cells so no aggregation necessary prior to reprojecting to Mollweide.

For each gear type, summarize total discards per cell; rasterize and save out to Mazu.  Note data span 2015-2017; taking all catch over these years may help smooth annual variation.

### Summarize bycatch (discards) per cell by gear type

```{r}

gear_short_df <- gear_clean_df %>%
  select(code = f_gear_code, name = fleet_gear_name) %>%
  distinct() %>%
  arrange(code)

### set up a base raster
xyz_df <- cell_df %>%
  select(x = lon_centre, y = lat_centre, z = cell)

cell_id_rast <- rasterFromXYZ(xyz_df, crs = '+init=epsg:4326') %>%
  extend(extent(c(-180, 180, -90, 90)), value = NA)

### Set up a filename stem
bycatch_fstem <- here_anx('stressors', 'fishing', 
                          'bycatch_by_gear', 'bycatch_%s_%s_sum.tif')
# unlink(list.files(dirname(bycatch_fstem), full.names = TRUE))
benthic_gears <- c('trawl', 'dredge', 'trap')

for(i in 1:nrow(gear_short_df)) {
  # i <- 1
  g_code <- gear_short_df$code[i]
  g_name <- gear_short_df$name[i]
  g_water_col <- ifelse(g_name %in% benthic_gears, 'benth', 'pel')
  
  bycatch_rast_f <- sprintf(bycatch_fstem, g_name, g_water_col)
  if(!file.exists(bycatch_rast_f)) {
    message('Processing ', bycatch_rast_f, '...')
    tmp_bycatch_df <- catch_gear_df %>%
      filter(f_gear_code == g_code) %>%
      group_by(cell) %>%
      summarize(discard_ind_sum  = sum(discards_ind, na.rm = TRUE),
                discard_nind_sum = sum(discards_nind)) %>%
      mutate(discard_sum = discard_ind_sum + discard_nind_sum)
    
    tmp_bycatch_rast <- raster::subs(cell_id_rast, tmp_bycatch_df,
                                     by = 'cell', which = 'discard_sum')
    writeRaster(tmp_bycatch_rast, bycatch_rast_f, overwrite = TRUE)
  } else {
    # message('File ', bycatch_rast_f, ' exists... skipping!')
  }
}
```

## Normalize bycatch by NPP

Rather than trying to use some variation on CPUE to estimate standing stock biomass, we will simply use NPP as prior CHI efforts have done.  This avoids issues of estimating non-targeted biomass based on effort on targeted stocks.  It will also be far easier and consistent than trying to use CPUE for targeted fishing effort later on.

Use NPP layers from Bio-ORACLE (https://www.bio-oracle.org/release-notes-2-2.php):

* BO2_ppmean_ss: Primary production (mean)
    * Mean sea surface net primary productivity of carbon, g/m^3/day
* BO2_ppmean_bdmean: Primary production (mean at mean depth)
    * Mean net primary productivity of carbon at mean bottom depth

> The data are documented in two peer reviewed articles that you should cite:
> 
> Tyberghein L, Verbruggen H, Pauly K, Troupin C, Mineur F, De Clerck O (2012) Bio-ORACLE: A global environmental dataset for marine species distribution modelling. Global Ecology and Biogeography, 21, 272–281.
> 
> Assis, J., Tyberghein, L., Bosh, S., Verbruggen, H., Serrão, E. A., & De Clerck, O. (2017). Bio-ORACLE v2.0: Extending marine data layers for bioclimatic modelling. Global Ecology and Biogeography.

Idea: use sea surface NPP to normalize pelagic/surface/midwater bycatch, and bottom concentration, combined with export flux from surface, to normalize benthic bycatch.

From Yool et al (2009), "Low efficiency of nutrient translocation for enhancing oceanic uptake of carbon dioxide": table 1, control, in Gt C a^-1^ (globally integrated):

| Field               | Control |
|:--------------------|--------:|
| Primary production  |  35.143 |
| Export flux,  100 m |   7.428 |
| Export flux,  200 m |   4.934 |
| Export flux,  500 m |   2.986 |
| Export flux, 1000 m |   1.715 |

From this, can we generate a profile of flux per depth? use non-linear least squares to estimate parameter $r, \theta$ for exponential decay:
$f(d) = f_0 \times (1 - r)^{d^{\theta}}$

```{r}
c_flux_df <- tribble(
  ~depth,  ~flux,
       0, 35.143,
     100,  7.428,
     200,  4.934,
     500,  2.986,
    1000,  1.715
) %>%
  mutate(f_rel = flux / max(flux))

flux_nls <- nls(f_rel ~ (1 - r) ^ (depth^theta), data = c_flux_df, start = list(r = 0.01, theta = 0.8))

r <- coef(flux_nls)[1]
th <- coef(flux_nls)[2]

flux_model <- c_flux_df %>%
  mutate(flux_pred = max(flux) * (1 - r)^(depth^th))

ggplot(flux_model, aes(x = depth, y = flux)) +
  geom_point(color = 'red3') +
  geom_line(color = 'red3') +
  geom_point(aes(y = flux_pred), color = 'green3') +
  geom_line(aes(y = flux_pred), color = 'green3') +
  theme_ohara() +
  scale_y_log10() +
  labs(y = 'Export flux, Gt C per year',
       x = 'Depth, m')
```

### Read and transform benthic and pelagic bycatch layers

The rasters of bycatch by gear are at 0.5° resolution, so no aggregation needed.

For each gear type:

* read in appropriate bycatch layers by pelagic or benthic gear type as a raster stack
* sum bycatch values across all gear types
* convert result to catch per km<sup>2</sup>
* reproject to 10 km Mollweide CRS (using nearest neighbor to avoid undue interpolation of the original data)
* convert to catch per 10 km x 10 km cell: bycatch/km<sup>2</sup> * 100 km<sup>2</sup>. 
    * Note that accounting for proportional ocean area of coastal cells, as is done elsewhere, in this case loses a lot of catch... likely an artifact due to total catch in large cells being attributed as uniformly distributed

```{r}
### proportional ocean area raster as raster base and mask
ocean_r_mol <- raster(here('_spatial/ocean_area_mol.tif'))

bycatch_fs <- list.files(here_anx('stressors', 'fishing', 'bycatch_by_gear'),
                         full.names = TRUE)
```

To do this we will need an ocean area raster.  The AquaMaps HCAF file has ocean area but it does not entirely agree with the NaturalEarth-generated rasters used in the rest of this analysis, so create one from scratch.

``` {r create ocean area raster at half deg res}
### set up ocean area raster: create blank at 0.01 degree
ocean_a_0.5_f <- here('_spatial/ocean_area_wgs84_0.5deg.tif')

if(!file.exists(ocean_a_0.5_f)) {
  r_0.01_base <- raster(ext = extent(c(-180, 180, -90, 90)),
                      res = 0.01, crs = '+init=epsg:4326')
  ocean_sf <- read_sf(here('_spatial/ne_10m_ocean/ne_10m_ocean.shp'))
  
  ocean_r_0.01 <- fasterize::fasterize(ocean_sf, r_0.01_base)
  ocean_a_0.01 <- area(r_0.01_base, progress = 'text') %>%
    mask(ocean_r_0.01, progress = 'text')
  
  ocean_a_0.5 <- aggregate(ocean_a_0.01, fact = 50, fun = 'sum', 
                           progress = 'text')
  
  writeRaster(ocean_a_0.5, ocean_a_0.5_f, overwrite = TRUE)
}

ocean_a_hcaf <- raster(ocean_a_0.5_f)
ocean_a_mol <- raster(here('_spatial/ocean_area_mol.tif'))
```

``` {r bycatch from pelagic gear types}
### read in pelagic gear types and sum
pel_sum_r <- stack(bycatch_fs[str_detect(bycatch_fs, '_pel_sum.tif')]) %>%
  calc(fun = sum, na.rm = TRUE)
# sum(values(pel_sum_r), na.rm = TRUE) # 4276031 total tonnes

### convert to bycatch per km^2
pel_sum_km2_r <- pel_sum_r / ocean_a_hcaf

### reproject
pel_sum_km2_r_moll <- pel_sum_km2_r %>%
  projectRaster(ocean_a_mol, method = 'ngb')

### convert to total catch
pel_r <- pel_sum_km2_r_moll * ocean_a_mol * 100
# sum(values(pel_r), na.rm = TRUE) 
# # 4375131, a little high but not terribly so

# get_qtiles(pel_r)
#          50%          90%          95%          99%        99.9%         100% 
# 5.415718e-02 1.379041e+00 7.574325e+00 4.797353e+01 6.078182e+02 2.696722e+04 
```

``` {r bycatch from benthic gear types}
ben_sum_r <- stack(bycatch_fs[str_detect(bycatch_fs, '_benth_sum.tif')]) %>%
  calc(fun = sum, na.rm = TRUE) 
# sum(values(ben_sum_r), na.rm = TRUE) # 40749287

### convert to bycatch / km^2
ben_sum_km2_r <- ben_sum_r / ocean_a_hcaf

### reproject
ben_sum_km2_r_moll <- ben_sum_km2_r %>%
  projectRaster(ocean_a_mol, method = 'ngb')

### convert to total catch
ben_r <- ben_sum_km2_r_moll * ocean_a_mol * 100
# sum(values(ben_r), na.rm = TRUE) 
# # 42127285, again a little high but not terribly so

# get_qtiles(ben_r)
#          50%          90%          95%          99%        99.9%         100% 
# 0.000000e+00 2.686037e+00 6.446003e+01 5.968119e+02 6.541804e+03 8.807373e+05 
```

### Set up surface NPP layer

For surface and midwater bycatch, estimate biomass based on surface productivity (log-transformed).  

Because this ius at a resolution comparable to 10 km cells (0.0833°), no need for aggregation prior to reprojecting.

Both the surface and benthic NPP layer contains some areas where no data is present, mostly in Arctic and Antarctic areas.  Use an iterative `focal()` process to fill out these layers (see `focal_gapfill()` function in `stressor_fxns.R` script)

```{r surface NPP}
surf_npp_f <- here_anx('stressors/fishing/npp', 'log_npp_present_surf_gf_mol.tif')

if(!file.exists(surf_npp_f)) {
  surface_npp_raw <- raster(here_anx('stressors/fishing/npp',
                                     'Present.Surface.Primary.productivity.Mean.tif'))
  # class      : RasterLayer 
  # dimensions : 2160, 4320, 9331200  (nrow, ncol, ncell)
  # resolution : 0.08333333, 0.08333333  (x, y)
  # extent     : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # source     : Present.Surface.Primary.productivity.Mean.tif 
  # names      : Present.Surface.Primary.productivity.Mean 
  # values     : 6e-05, 0.257881  (min, max)

  surface_npp_mol <- projectRaster(surface_npp_raw, ocean_r)
  
  log_surf_npp_mol <- log_plus(surface_npp_mol)
  
  log_surf_npp_gapfill <- focal_gapfill(log_surf_npp_mol, iters = 50)
  # x <- check_missing_cells(log_surf_npp_gapfill)
  # sum(!is.na(values(x)))
  # a few antarctic cells still NA - ice sheets etc?
  # y <- mask(pel_r, x); sum(values(y), na.rm = TRUE)
  # z <- mask(ben_r, x); sum(values(z), na.rm = TRUE)
  
  # get_qtiles(log_surf_npp_gapfill)
  #       50%       90%       95%       99%     99.9%      100% 
  # 0.7790794 1.5680076 1.8177607 2.4418932 3.3707904 4.6151205
  writeRaster(log_surf_npp_gapfill, surf_npp_f, overwrite = TRUE)
} else {
  log_surf_npp_gapfill <- raster(surf_npp_f)
}
```

### Set up benthic NPP/surface flux layer

For benthic bycatch, we will normalize by a combination of surface NPP, as export flux per the model above, and benthic NPP where non-zero.  First create a layer that converts depth to the $(1 - r)^{d^{\theta}}$ portion of our flux model:

$flux(depth) = NPP_{surface} \times (1 - r)^{d^{\theta}}$

where $r = 0.3412355$ and $\theta = 0.2877847$.

```{r create depth raster}
flux_r_f <- here_anx('stressors/fishing/npp', 'flux_model.tif')

flux_model <- function(d) {
  ### function to calculate flux based on depth
  values(d)[values(d) > 1000] <- 1000 ### cap depth adjustment at 1000 m value
  f <- (1 - 0.3412355)^(d^0.2877847)
}

if(!file.exists(flux_r_f)) {
  ### read in surface NPP as base raster
  surface_npp_raw <- raster(here_anx('stressors/fishing/npp',
                                     'Present.Surface.Primary.productivity.Mean.tif'))
  
  gebco_rast_file <- here_anx('_spatial/gebco/GRIDONE_1D.nc')
  elev_rast  <- raster(gebco_rast_file)
  # class      : RasterLayer 
  # dimensions : 10801, 21601, 233312401  (nrow, ncol, ncell)
  # resolution : 0.01666667, 0.01666667  (x, y)
  # extent     : -180.0083, 180.0083, -90.00833, 90.00833  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # source     : GRIDONE_1D.nc 
  # names      : layer 
  # zvar       : z 
  ### The GEBCO raster is elevation, so depth is neg... mult by -1, clip out
  ### cells with neg depth, aggregate by a factor of 5 to NPP res using mean 
  ### depth
  depth_rast <- elev_rast * -1
  values(depth_rast)[values(depth_rast) < 0] <- NA
  depth_rast_agg <- aggregate(depth_rast, 
                              fact = 5, fun = mean)
  ### even at same res, different extents etc.
  depth_rast_npp <- depth_rast_agg %>%
    projectRaster(surface_npp_raw, ### target CRS/res/extent
                  progress = 'text',
                  method = 'bilinear')
  
  flux_rast <- flux_model(depth_rast_npp)
  
  writeRaster(flux_rast, flux_r_f, overwrite = TRUE)
}
```

``` {r surface export flux plus benthic NPP}
benth_npp_f <- here_anx('stressors/fishing/npp', 'log_npp_present_benth_gf_mol.tif')
# unlink(benth_npp_f)
if(!file.exists(benth_npp_f)) {
  benthic_npp_raw <- raster(here_anx('stressors/fishing/npp',
                                     'Present.Benthic.Mean.Depth.Primary.productivity.Mean.tif'))
  # class      : RasterLayer 
  # dimensions : 2160, 4320, 9331200  (nrow, ncol, ncell)
  # resolution : 0.08333333, 0.08333333  (x, y)
  # extent     : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # source     : Present.Benthic.Mean.Depth.Primary.productivity.Mean.tif 
  # names      : Present.Benthic.Mean.Depth.Primary.productivity.Mean 
  # values     : 0, 0.376071  (min, max)
  
  surface_npp_raw <- raster(here_anx('stressors/fishing/npp',
                                     'Present.Surface.Primary.productivity.Mean.tif'))
  flux_model_rast <- raster(flux_r_f)
  
  export_flux_rast <- surface_npp_raw * flux_model_rast
  
  ### sum benthic NPP and export flux
  benth_total_rast <- benthic_npp_raw + export_flux_rast
  
  benth_total_moll <- benth_total_rast %>%
    projectRaster(ocean_r)
  log_ben_total_C_mol <- log_plus(benth_total_moll)
  
  log_ben_tot_npp_gapfill <- focal_gapfill(log_ben_total_C_mol, iters = 50)
  # x <- check_missing_cells(log_ben_tot_npp_gapfill)
  
  # get_qtiles(log_ben_tot_npp_gapfill)
  #      50%      90%      95%      99%    99.9%     100% 
  # 0.420190 1.021973 1.227471 1.947043 2.844674 4.615121
  
  writeRaster(log_ben_tot_npp_gapfill, benth_npp_f, overwrite = TRUE)
} else {
  log_ben_tot_npp_gapfill <- raster(benth_npp_f)
}
```

### Normalize bycatch and write out

Normalize the pelagic and benthic bycatch values (untransformed) by the appropriate NPP layer (log-transformed) then rescale, using a 99.9%ile as reference point.  These new layers seem substantially less right-skewed (though still very very skewed) than the ones from CHI Recent Pace of Change project.

```{r finalize pelagic bycatch}
pel_normalized_r <- pel_r / log_surf_npp_gapfill
# get_qtiles(pel_normalized_r)
#          50%          90%          95%          99%        99.9%         100% 
# 8.342141e-02 2.428338e+00 6.178333e+00 2.758328e+01 1.463990e+02 1.441491e+04 

pel_rescaled_r <- rescale_stressor(pel_normalized_r, qtile = .999)
# get_qtiles(pel_rescaled_r)
#          50%          90%          95%          99%        99.9%         100% 
# 0.0005698224 0.0165871279 0.0422020262 0.1884116881 0.9999395458 1.0000000000 

### compare to old pelagic high bycatch stressor (~1 km Mollweide)
# dir <- '/home/shares/ohi/git-annex/impact_acceleration/stressors/comm_fish/final'
# r <- raster(file.path(dir, 'pel_hb', 'pel_hb_2014_rescaled_mol.tif'))
# get_qtiles(r)
#          50%          90%          95%          99%        99.9%         100% 
# 9.386958e-08 3.750110e-04 7.535241e-03 5.118972e-02 1.776541e-01 1.000000e+00

plot(pel_rescaled_r, axes = FALSE, main = 'Pelagic bycatch stressor',
     col = hcl.colors(20))
plot(log10(pel_rescaled_r), axes = FALSE, main = 'Pelagic bycatch log10(stressor)',
     col = hcl.colors(20))

hist(pel_rescaled_r, main = 'Pelagic bycatch stressor')
hist(log10(pel_rescaled_r), main = 'Pelagic bycatch log10(stressor)')

writeRaster(pel_rescaled_r, 
            here('_data/stressors_mol/bycatch_pelagic_2017.tif'),
            overwrite = TRUE)
```

``` {r finalize benthic bycatch}
ben_normalized_r <- ben_r / log_ben_tot_npp_gapfill
# get_qtiles(ben_normalized_r)
#          50%          90%          95%          99%        99.9%         100% 
# 0.000000e+00 8.892644e+00 1.126888e+02 1.168168e+03 9.867294e+03 2.135482e+07 

ben_rescaled_r <- rescale_stressor(ben_normalized_r, qtile = .999)
# get_qtiles(ben_rescaled_r)
#          50%          90%          95%          99%        99.9%         100% 
# 0.0000000000 0.0009012241 0.0114204333 0.1183878551 0.9998903073 1.0000000000 

### compare to old pelagic high bycatch stressor
# r1 <- raster(file.path(dir, 'dem_nondest_hb', 'dem_nondest_hb_2014_rescaled_mol.tif'))
# r2 <- raster(file.path(dir, 'dem_dest', 'dem_dest_2014_rescaled_mol.tif'))
# get_qtiles(r1)
#          50%          90%          95%          99%        99.9%         100% 
# 0.000000e+00 9.992018e-06 9.858453e-04 2.784128e-02 2.191673e-01 1.000000e+00
# get_qtiles(r2)


plot(ben_rescaled_r, axes = FALSE, main = 'Benthic bycatch stressor',
     col = hcl.colors(20))
plot(log10(ben_rescaled_r), axes = FALSE, main = 'Benthic bycatch log10(stressor)',
     col = hcl.colors(20))
hist(ben_rescaled_r, main = 'Benthic bycatch stressor')
hist(log10(ben_rescaled_r), main = 'Benthic bycatch log10(stressor)')

writeRaster(ben_rescaled_r, 
            here('_data/stressors_mol/bycatch_benthic_2017.tif'),
            overwrite = TRUE)

```
