==> 1_setup/stressors/4c_targeted_fishing_normalize_npp.Rmd <==
---
title: "Stressors: targeted fishing by species and cell - finalize stressors per spp"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(oharac) ### remotes::install_github('oharac/oharac')
  ### includes some helper functions, e.g., dt_join
library(tidyverse)
library(here)

source(here('common_fxns.R'))
source('stressor_fxns.R') ### in same dir as this Rmd
```

# Summary

Read in rasters (as csvs) of fisheries catch per species per cell.  Normalize by NPP (surface and benthic); identify a reference point; save as .csvs of cell ID and normalized catch. 

# Data

Data from

* Watson, R.A. and Tidd, A.N. (2018) Mapping nearly a century and a half of global marine fishing: 1869 to 2015. Marine Policy 93, 171-177
* Watson, R. (2017) A database of global marine commercial, small-scale, illegal and unreported fisheries catch 1950-2014. Nature Scientific Data 4 (170039).

Species distributions from

* AquaMaps
* IUCN Red List

# Methods

Briefly, here is the process for finalizing catch stressor layers:

* Process catch data for the species:
    * Read in catch per LOICZID cell (for both pelagic and benthic gears)
    * Convert catch per 0.5° cell to catch per km^2^
    * Reproject to Mollweide (either projectRaster or join to a lookup dataframe of LOICZID to Mollweide cell ID)
    * Convert to pelagic and benthic catch per 10 x 10 km cell.
* Divide each processed catch layer (benthic and pelagic) by the appropriate NPP layer (log-transformed)
* Convert to dataframe of cell ID and NPP-normalized catch, and write out.

## Examine NPP layers

Can these be aggregated upward to 0.5 degrees with little loss of information?  Examine mean and variance of aggregated results.  Low variance within an aggregated cell means little information is lost in the aggregation process.

```{r, eval = FALSE}
surface_npp_raw <- raster(here_anx('stressors/fishing/npp',
                                   'Present.Surface.Primary.productivity.Mean.tif'))
# class      : RasterLayer 
# dimensions : 2160, 4320, 9331200  (nrow, ncol, ncell)
# resolution : 0.08333333, 0.08333333  (x, y)
# extent     : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
# crs        : +proj=longlat +datum=WGS84 +no_defs 
# source     : Present.Surface.Primary.productivity.Mean.tif 
# names      : Present.Surface.Primary.productivity.Mean 
# values     : 6e-05, 0.257881  (min, max)
agg_mean <- aggregate(surface_npp_raw, fact = 6, fun = mean, progress = 'text')
agg_sd   <- aggregate(surface_npp_raw, fact = 6, fun = sd, progress = 'text')
# agg_min  <- aggregate(surface_npp_raw, fact = 6, fun = min, progress = 'text')
# agg_max  <- aggregate(surface_npp_raw, fact = 6, fun = max, progress = 'text')

# agg_range <- agg_max - agg_min
agg_cv <- agg_sd / agg_mean
high_cv <- agg_cv; high_cv[high_cv < 0.1] <- NA

plot(agg_sd, axes = FALSE, col = hcl.colors(20), main = 'std dev')
plot(high_cv, axes = FALSE, col = hcl.colors(20), main = 'coef of var ≥ 0.1')

```
The coastal regions have a relatively high variance, as could be expected.  Because of this, we should probably retain the original resolution.

### Read in processed NPP layers

These layers are created in script `3_process_fishing_bycatch.Rmd`.  The surface NPP raster is the log(x + max(x)/100) transformed surface NPP, normalized by the 99.9th percentile.  The benthic NPP raster is the log-transformed combination of benthic NPP (where non-zero) and export flux from surface NPP, to account for the fact that particularly for deep benthic environments, food is derived from NPP sinking from the surface.  See the indicated script for the derivation of the export flux model.

```{r}
surf_npp  <- raster(here_anx('stressors/fishing/npp', 'log_npp_present_surf_gf_mol.tif'))
benth_npp <- raster(here_anx('stressors/fishing/npp', 'log_npp_present_benth_gf_mol.tif'))

plot(surf_npp, axes = FALSE, main = 'Surface log(NPP)', col = hcl.colors(n = 20))
plot(benth_npp, axes = FALSE, main = 'Benthic log(NPP)', col = hcl.colors(n = 20))
```



### Identify species position in water column

Depending on the gear type we will normalize catch by either surface or benthic NPP.

```{r}
watson_maps_lookup <- read_csv(here('_setup/stressors/int/watson_maps_lookup.csv'))
spp_vec <- watson_maps_lookup$spp %>% unique()

wcol_pos <- read_csv(here('_data/grouping_traits_post_imputation.csv')) %>%
  dplyr::select(species, wcol)

table(wcol_pos$wcol)
#   ben    bp   pel    rf 
# 28453  3832  4243 14804
### the vast majority of species are not pelagic (92%).  But from script 4b,
### the bulk of catch is not benthic.

# catch_fs <- list.files(here_anx('stressors/fishing/2_total_catch_by_spp_cell'), full.names = TRUE)
# 
# tot_catch_df <- parallel::mclapply(catch_fs, mc.cores = 20,
#                        FUN = function(f) {
#                          # f <- catch_fs[3]
#                          nm <- str_remove_all(basename(f), 'spp_catch_|.csv') %>%
#                            str_replace_all('[^a-z]+', ' ')
#                          x <- read_csv(f, show_col_types = FALSE)
#                          y <- x %>% summarize(pel_sum = sum(pel_tot_catch),
#                                               ben_sum = sum(ben_tot_catch)) %>%
#                            mutate(spp = nm)
#                        }) %>%
#   bind_rows()
# 
# benth_check <- tot_catch_df %>%
#   left_join(wcol_pos, by = c('spp' = 'species')) %>%
#   group_by(wcol) %>%
#   summarize(ben_tot = sum(ben_sum, na.rm = TRUE),
#             pel_tot = sum(pel_sum, na.rm = TRUE))
# 
# knitr::kable(benth_check)
```

Based on sample of about 16000 spp:

|water col pos | benthic total | pelagic total |
|:-------------|--------------:|--------------:|
|benthic       |   30,861,033  |   19,139,614  |
|benthopelagic |    7,720,392  |   23,181,578  |
|pelagic       |   10,930,387  |   75,876,780  |
|reef          |   13,033,919  |   22,844,775  |
|NA            |   11,217,167  |   19,549,369  |

Catch by benthic and pelagic gear types does not seem well linked to position in water column...

## Process catch layers

* Process catch data for the species:
    * Read in catch per LOICZID cell
    * Convert catch per 0.5° cell to catch per km^2^
    * Reproject to Mollweide (either projectRaster or join to a lookup dataframe of LOICZID to Mollweide cell ID)
    * Convert to catch per 10 x 10 km cell.
* Divide the processed catch layer by the appropriate NPP layer (log-transformed)
* Convert to dataframe of cell ID and NPP-normalized catch, and write out.

```{r set up dataframe of cell_id to loiczid to ocean_a lookup}
loiczid_mol <- raster(here('_spatial/loiczid_mol.tif'))
ocean_a_mol <- raster(here('_spatial/ocean_area_mol.tif'))

cell_id_df <- data.frame(loiczid = values(loiczid_mol),
                         ocean_a_mol = values(ocean_a_mol),
                         cell_id = 1:ncell(ocean_a_mol),
                         pel_npp = values(surf_npp),
                         ben_npp = values(benth_npp))

### using ocean area at half deg generated from Natural Earth
# ocean_a_0.5 <- raster(here('_spatial/ocean_area_wgs84_0.5deg.tif'))
# ocean_a_hcaf_df <- data.frame(ocean_a_hcaf = values(ocean_a_0.5),
#                               loiczid = 1:ncell(ocean_a_0.5))
```

```{r}
### to match species names with IUCN SIDs for grabbing all rangemaps:
iucn_spp_match <- read_csv(here('_data/iucn_spp/iucn_to_worms_match.csv')) %>%
  filter(mapped == TRUE)
get_spp_rangemap <- function(s, src) {
  ### s <- 'pristis pristis'
  rangemap_stem <- here_anx('spp_maps_mol/%s_spp_mol_%s.csv')
  
  if(src == 'iucn') { ### collect iucn_sid(s) for IUCN spp filename
    ids <- iucn_spp_match %>%
      filter(worms_name == s) %>%
      pull(iucn_sid)
    rangemap_f <- sprintf(rangemap_stem, src, ids)
    
  } else { ### use spp name for AM filename
    rangemap_f <- sprintf(rangemap_stem, src, str_replace_all(s, ' ', '_'))
  }
  
  ### filter presence or prob depending on which file type
  if(length(rangemap_f) > 1) {
    rangemap <- lapply(rangemap_f, FUN = data.table::fread) %>%
      data.table::rbindlist()
  } else {
    rangemap <- data.table::fread(rangemap_f)
  }
  if(src == 'am') {
    rangemap <- rangemap %>% filter(prob >= 0.50) %>%
      select(-prob) %>%
      distinct()
  } else {
    rangemap <- rangemap %>% filter(presence != 5) %>%
      select(-presence) %>%
      distinct()
  }
  return(rangemap)
}

process_species_to_moll <- function(f) {
  ### 14 30 46 62 78 94
  # f <- spp_hcaf_fs[14]
  # f <- spp_hcaf_fs[str_detect(spp_hcaf_fs, 'iucn_')][1]
  # f <- "/home/shares/ohi/spp_vuln/spp_vuln_mapping/stressors/fishing/2_total_catch_by_spp_cell/am_spp_catch_acanthomenia_arcuata.csv"
  new_f <- str_replace(basename(f), 'spp_catch', 'spp_npp_catch') %>%
    str_replace('.csv', '_mol.csv')
  spp_mol_map_f <- file.path(spp_mol_map_dir, new_f)
  
  s <- str_remove_all(basename(f), '.+_spp_catch_|.csv') %>%
    str_replace_all('_', ' ')
  src <- str_remove(basename(f), '_spp_catch.+')
  
  if(!file.exists(spp_mol_map_f)) {
    
    i <- which(spp_hcaf_fs == f)
    message('Processing NPP-norm spp-total catch map for ', s, 
            ' (', i, ' of ', length(spp_hcaf_fs), ') based on ',
            toupper(src), ' spp map...')

    ### Read in catch per LOICZID cell and species range map; check for missing
    spp_catch_hcaf <- data.table::fread(f)
    if(nrow(spp_catch_hcaf) == 1 & spp_catch_hcaf$loiczid == -1) {
      message('No valid catch map for ', s, ' in ', src, '...')
      return(NULL)
    }
        
    spp_rangemap <- get_spp_rangemap(s, src)
    if(nrow(spp_rangemap) == 0) {
      message('No valid range map for ', s, ' in ', src, '...')
      return(NULL)
    }
    
    ### "Reproject" to Mollweide by joining to LOICZID-cell_id lookup
    ### * calc area of LOICZID cell intersected with spp range
    ### * distribute total LOICZID catch among mollweide cells,
    ###   uniformly distributed across area
    spp_catch_mol <- spp_catch_hcaf %>%
      oharac::dt_join(cell_id_df, by = 'loiczid', type = 'left') %>%
      ### drop non-range cells:
      filter(cell_id %in% spp_rangemap$cell_id) %>%
      ### calc loiczid intersected with range:
      group_by(loiczid) %>%
      mutate(ocean_a_hcaf = sum(ocean_a_mol * 100)) %>%
      ungroup() %>%
      ### distribute total catch of LOICZID and normalize by NPP
      mutate(pel_sum = pel_tot_catch / ocean_a_hcaf * ocean_a_mol * 100,
             ben_sum = ben_tot_catch / ocean_a_hcaf * ocean_a_mol * 100,
             pel_norm = pel_sum / pel_npp,
             ben_norm = ben_sum / ben_npp) %>%
      drop_na()
    if(nrow(spp_catch_mol) == 0) {
      warning('No catch/spp range intersection: ', s, ' mapped in ', src, '...')
      return(NULL)
    }
    
    out_df <- spp_catch_mol %>%
      dplyr::select(cell_id, 
             pel_tot_catch = pel_sum,
             ben_tot_catch = ben_sum,
             pel_norm_catch = pel_norm,
             ben_norm_catch = ben_norm) %>%
      mutate(across(where(is.numeric), ~round(.x, 5)))
    
    write_csv(out_df, spp_mol_map_f)
  } else {
    # message('File ', basename(spp_mol_map_f), ' exists... skipping!')
  }
}
```

``` {r}
spp_hcaf_map_dir <- here_anx('stressors/fishing/2_total_catch_by_spp_cell')
spp_mol_map_dir  <- here_anx('stressors/fishing/3_npp_norm_catch_by_spp_moll')

spp_hcaf_fs <- list.files(spp_hcaf_map_dir, full.names = TRUE) %>% unique()

### to cut zero-length files
# x <- list.files(spp_mol_map_dir, full.names = TRUE, pattern = 'am_spp_npp_catch_')
# df <- data.frame(f = x, fsize = file.size(x))
# y <- df %>% filter(fsize < 100) %>% .$f
# unlink(y)

### to run only missing files
x <- list.files(spp_mol_map_dir)
y <- x %>% str_replace('spp_npp_catch', 'spp_catch') %>% str_replace('_mol.csv', '.csv')
z <- file.path(spp_hcaf_map_dir, y)
spp_hcaf_fs <- spp_hcaf_fs[!spp_hcaf_fs %in% z]

### to clear all files and rerun:
# zxcv <- list.files(spp_mol_map_dir, full.names = TRUE)
# unlink(zxcv)

tmp <- parallel::mclapply(spp_hcaf_fs, mc.cores = 32, FUN = process_species_to_moll)
### try_check <- sapply(tmp, class) %>% unlist()
### which(try_check == 'try-error')

### All "missing" files are due to lack of spp map &/or catch map
```
