==> 1_setup/stressors/7_process_other_stressors.Rmd <==
---
title: "Gather stressors and process to analysis CRS"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(sf)
library(tidyverse)
library(here)
library(RColorBrewer)

source(here('common_fxns.R'))
source('stressor_fxns.R') ### must be in same directory for relative path

```

# Summary

Gather stressor layers and process to the CRS for this analysis: Mollweide, 10 km x 10 km.  Identify gapfill needs if necessary.  Layers gathered/processed here:

* Benthic disturbance (from effort in destructive fishing practices)
* Direct human (from human population density)
* Extreme temperature events
* Sea Level Rise
* Ocean Acidification
* Ultraviolet
* Light pollution

Note that for some stressors, we may need to convert a raw value (e.g., ship track count) to an intensity (count per km^2) prior to reprojecting.

Note in methodology: where in CHI at ~1 km Mollweide we frequently use the 99.99%ile as a reference point, here I am using the 99.9%ile (one fewer decimal).  Since we've aggregated by a factor of 100 (10 x 10), it seems like 99.9% will still clip out the very most extreme values.



# Data

All files are located on Mazu, starting from `/home/shares/ohi/` unless otherwise noted.

## Benthic disturbance from fishing

* GFW data to estimate benthic disturbance from food systems project:   
    * dir: `stressors_2021/_dataprep/fishing_benthic_destructive`
        * file: `FishingWatch_annual_effort_destructive_hours_2017.tif`
        * log-transform
    * cite food systems project where GFW data was processed to this layer?
    * or cite GFW and include methods/code to generate this layer within this project.
    * Information from the food systems project: https://github.com/OHI-Science/food_systems/tree/master/fisheries/marine/disturbance

## Direct human impact/habitat destruction 

using population as proxy:

* dir: `git-annex/globalprep/mar_prs_population/v2020/int/`
    * file: `human_density_2020_mol_tmp.tif`
    * also files for `human_count...` and `yearly_change...`
    * Log-transform

## Organic chemical pollution?

## Microplastics

Using data from van Sebille (2015):

* Van Sebille, E., Wilcox, C., Lebreton, L., Maximenko, N., Hardesty, B. D., Van Franeker, J. A., Eriksen, M., Siegel, D., Galgani, F., & Law, K. L. (2015). A global inventory of small floating plastic debris. Environmental Research Letters, 10(12), 124006.
* dir: `stressors/microplastics`
    * file: `vansebille_mass.csv`
        

## Climate stressors other than SST

Count these as intensities - no conversion to intensity required.

### Extreme temperature events

These maps were created for Mel's MARISCO stressors project.  See there for methodology.  Nutshell:

1. For each week and raster cell: determine the anomaly value (observed weekly temp - climatological weekly mean) that corresponds to the 90th quantile using years ~1985-2015.
2. Then for each week/raster cell: determine the weeks in each year of 2016-2020 that exceed the 90th quantile value calculated using the 1985-2015 data.
3. Sum exceedences occurring from 2016-2020.

* Use annual mean SLR anomaly rasters from here:
    * dir: `/stressors_2021/_dataprep/SST_past_present/final_extreme`
        * file: `sst_exceeds_90quant_2016-2020.tif`
    * These are projected to Mollweide, but not rescaled 0-1 (currently, sum of extreme weeks / total number of weeks)
* See this issue for methods, for now: https://github.com/mapping-marine-spp-vuln/issues/issues/22

### Sea level rise from OHI 2021

* Use annual mean SLR anomaly rasters from here:
    * dir: `git-annex/globalprep/prs_slr/v2021/int/msla_annual_mean/`
        * file: `msla_annual_2019.tif`
    * These are pre-transform and mask.  Transformed and masked are also available, but let's use these so we can make sure the transform and mask are accurate to our particular CRS.
* See OHI processing script for details: `https://github.com/OHI-Science/ohiprep_v2021/blob/gh-pages/globalprep/prs_slr/v2021/slr_layer_prep_v2.Rmd`
    
### Ocean acidification from CHI Pace of Change (2017 data year)

* dir: `git-annex/impact_acceleration/stressors/oa/final`
    * file: `oa_2017_rescaled_mol.tif`
* These are in Mollweide but at 934 meter resolution... reproject to 10 km
    
### Ultraviolet radiation from OHI 2021

* dir: `git-annex/globalprep/prs_uv/v2021/int/rescaled/`
    * file: `annual_diff_2016_2020_rescaled.tif`
* See OHI processing script for details: `https://ohi-science.org/ohiprep_v2020/globalprep/prs_uv/v2020/uv_dataprep.html`

## Light pollution

* Light pollution from [Li et al. (2020) A harmonized global nighttime light dataset 1992–2018](https://www.nature.com/articles/s41597-020-0510-y#Sec8)
    * Night time light averaged over year, at 30 arc-second resolution.
    * Downloaded data from FigShare 11/30/21: https://figshare.com/articles/dataset/Harmonization_of_DMSP_and_VIIRS_nighttime_light_data_from_1992-2018_at_the_global_scale/9828827/2?file=17626049

# Methods

For each stressor layer, reproject to 10 km Mollweide CRS and mask to ocean cells.

From `?projectRaster`: 

> Note:
> If the resolution of the output is much larger than that of the input, you should first aggregate the input such that the resolution of the input becomes more similar (perhaps a little smaller) to the output.

```{r define stuff}
here_ohi <- function(f = '', ...) {
  f <- paste(f, ..., sep = '/')
  f <- stringr::str_replace_all(f, '\\/+', '/')
  sprintf('/home/shares/ohi/%s', f)
}

### Load ocean proportion in Mollweide 10 km x 10 km (analysis CRS and res)
ocean_rast <- raster(here('_spatial/ocean_area_mol.tif'))

### ocean area (in km^2) in 10 km x 10 km Mollweide
ocean_area_mol <- ocean_rast * 100

### load ocean area (in km^2) in 0.1 deg WGS84 (for converting values to 
### intensities prior to reprojecting)
ocean_area_0.1deg <- raster(here('_spatial/ocean_area_wgs84_0.1deg.tif'))

### Code to create ocean area raster at 0.1 degree:                            
# base_rast_0.01deg <- raster(ext = extent(c(-180, 180, -90, 90)), 
#                            res = 0.01, crs = '+init=epsg:4326')
# ocean_mask_sf <- read_sf(here('_spatial/ne_10m_ocean/ne_10m_ocean.shp'))
# ocean_rast_0.01deg <- fasterize::fasterize(ocean_mask_sf, base_rast_0.01deg)
# ocean_rast_0.1deg <- aggregate(ocean_rast_0.01deg, 
#                                progress = 'text',
#                                fact = 10, fun = sum)
# ocean_rast_0.1deg <- ocean_rast_0.1deg / 100
# ocean_area_0.1deg <- ocean_rast_0.1deg * raster::area(ocean_rast_0.1deg)
# writeRaster(ocean_area_0.1deg, here('_spatial/ocean_area_wgs84_0.1deg.tif'))
```

## Global Fishing Watch for benthic disturbance

Reproject, then rescale 0-1.

GFW data for OHI at 0.01 degree resolution.  Aggregate by a factor of about 10x to get closer to 10 km x 10 km target resolution (at equator...), convert to intensity (hours/km^2^) then log transform and reproject.


```{r GFW data for benthic disturbance}
benth_dist_f <- here('_data/stressors_mol/fishing_benthic_dest_2017.tif')
### unlink(benth_dist_f)

if(!file.exists(benth_dist_f)) {
  local_f <- here_anx('stressors/fishing',
                     'FishingWatch_annual_effort_destructive_hours_2017.tif')
  if(!file.exists(local_f)) {
    orig_f <- here_ohi('stressors_2021/_dataprep', 
                       'fishing_benthic_destructive', 
                       'FishingWatch_annual_effort_destructive_hours_2017.tif')
    file.copy(orig_f, local_f)
  }
  raw_r <- raster(local_f)
  # class      : RasterLayer 
  # dimensions : 18000, 36000, 6.48e+08  (nrow, ncol, ncell)
  # resolution : 0.01, 0.01  (x, y)
  # extent     : -180.005, 179.995, -89.995, 90.005  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # source     : FishingWatch_annual_effort_destructive_hours_2017.tif 
  # names      : FishingWatch_annual_effort_destructive_hours_2017 
  # values     : 0, 1540.895  (min, max)
  
  ### Aggregate (sum for total hours in cell) and convert to intensity - hours per km^2
  agg_r <- aggregate(raw_r, fact = 10, fun = 'sum', progress = 'text')
  intensity_r <- agg_r / ocean_area_0.1deg
    
  ### reproject intensity to Mollweide and clip out tiny neg numbers
  reproj_intens_r <- projectRaster(intensity_r, ocean_rast, progress = 'text')
  reproj_intens_r[reproj_intens_r < 0] <- 0
  
  ### rescale the reprojected raster, fill all NAs with 0, then mask to just ocean
  reproj_log_r <- log_plus(reproj_intens_r)
  reproj_log_r <- rescale_stressor(reproj_log_r)
  values(reproj_log_r)[is.na(values(reproj_log_r))] <- 0

  log_r_ocean <- mask(reproj_log_r, ocean_rast, progress = 'text')
  
  ### round to drop ludicrous precision then save out
  values(log_r_ocean) <- round(values(log_r_ocean), 5)
  writeRaster(log_r_ocean, benth_dist_f, overwrite = TRUE)
  
}

r <- raster(benth_dist_f)
  
plot(r, axes = FALSE, 
     main = 'rescaled, log(x+xmax/1000)-transformed annual destructive hours')

hist(r[r > 0], main = 'distribution of non-zero values')

```

## Land-based stressors

### Direct human impact/habitat destruction using population as proxy

The data for this is population density or count (on land).  We want an indication of the number of people living within 25 km (or so) of the coast.  To do this, try `raster::focal()`: Calculate focal ("moving window") values for the neighborhood of focal cells using a matrix of weights, perhaps in combination with a function.

* To get a circular window on the 934 m native resolution of OHI population data, use a matrix with 55 rows and columns.  Assign value of 1 for cells within a 27-cell radius of the center.
* Fill all NAs with zeros, then apply `focal()` to sum up the total population within the circular window, and fill the cell with that value.
* Then mask to just coastal cells, which have now been filled with population within 25 km.  As the cells edge out from shore, the farther cells will have lower values, and aggregating/reprojecting to 10 km will be affected by this as it will bring down the average; but since we are rescaling the results from 0 to 1, this should not be a critical concern.

Rescale based on 99.9%ile, without transformation.  Seems like a linear relationship between number of people and their impact makes sense.

```{r Population data}
pop_f <- here('_data/stressors_mol/direct_human_2020.tif')
### unlink(pop_f)

if(!file.exists(pop_f)) {
  local_f <- here_anx('stressors/population',
                      'human_density_2020_mol.tif')
  if(file.exists(local_f)) {
    orig_f <- here_ohi('git-annex/globalprep/mar_prs_population/v2020/int/',
                       'human_density_2020_mol.tif')
    file.copy(orig_f, local_f)
  }
  raw_r <- raster(orig_f)
  # class      : RasterLayer 
  # dimensions : 19305, 38610, 745366050  (nrow, ncol, ncell)
  # resolution : 934.4789, 934.4789  (x, y)
  # extent     : -18040095, 18040134, -9020047, 9020067  (xmin, xmax, ymin, ymax)
  # crs        : +proj=moll +lon_0=0 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 
  # source     : human_density_2020_mol.tif 
  # names      : human_density_2020_mol 
  # values     : 0, 101691.8  (min, max)
  
  pop_r <- raw_r
  values(pop_r)[is.na(values(pop_r))] <- 0
  
  focal_pop_f <- here_anx('stressors/population/pop_focal_25km_radius_1km_mol.tif')
  if(!file.exists(focal_pop_f)) {
    ### set up focal matrix - weight of 1 within 25 km, 0 outside
    focal_rad <- 27
    focal_dia <- focal_rad*2+1
    focal_mtx <- matrix(nrow = focal_dia, ncol = focal_dia)
    for(i in 1:focal_dia) {
      for(j in 1:focal_dia) {
        ### i <- 1; j <- 1
        r_from_center <- sqrt((i - (focal_rad + 1))^2 + (j - (focal_rad + 1))^2)
        focal_mtx[i, j] <- ifelse(r_from_center < focal_rad, 1, 0)
      }
    }
    ### Calculate focal sum of population (or pop density, for equal area
    ### there is no meaningful difference when rescaling anyway)
    focal_pop_r <- focal(pop_r, w = focal_mtx, fun = sum, na.rm = TRUE, 
                         progress = 'text')
    
    writeRaster(focal_pop_r, focal_pop_f, overwrite = TRUE)
  }
  
  focal_pop_r <- raster(focal_pop_f)
  
  ### Already in Mollweide, but wrong resolution... log transform, aggregate 
  ### up to ~ 10 km, then reproject
  log_focal_pop_r <- log_plus(focal_pop_r)
  agg_focal_pop_r <- aggregate(log_focal_pop_r, fact = 10, fun = 'mean', progress = 'text')
  agg_focal_pop_r_10km <- projectRaster(agg_focal_pop_r, ocean_rast, progress = 'text')
  
  ### Masking to coastal cells rather than whole ocean, then
  ### rescale based on cropped raster to 99.9%ile
  coastal_r <- raster(here('_spatial/coastal_3nmi_area_mol.tif'))
  pop_coastal <- mask(agg_focal_pop_r_10km, coastal_r, progress = 'text') %>%
    rescale_stressor(qtile = .999)

  ### round to drop ludicrous precision then save out
  values(pop_coastal) <- round(values(pop_coastal), 5)
  writeRaster(pop_coastal, pop_f)
  
}

r <- raster(pop_f)
  
plot(r, axes = FALSE,
     main = 'Human population density: log(x+1), rescaled to 99.9%ile')

hist(r[r > 0], main = 'distribution of non-zero values')

```

### Chemical pollution?

### Microplastic pollution

The .csv contains microplastic mass based on Erik van Sebille's model in 2015.  Columns are longitude degrees 0 to 360, and rows are latitude 90 to -90.  Column 0 and 360 should be the same but do not agree...? set 0 column to be average of 0 and 360 columns.  Align to cell centers, presuming that coords are lower left corner of cell (checked via visual fit of filled cells against known global map).

```{r microplastics}
microplastics_f <- here('_data/stressors_mol/microplastics_2015.tif')
### unlink(microplastics_f)

if(!file.exists(microplastics_f)) {
  data_f <- here_anx('stressors/microplastics',
                     'vansebillemodel_mass.csv')
  raw_df <- read_csv(data_f, col_names = FALSE, col_types = cols(.default = 'd')) 
  
  check_0_360 <- raw_df %>%
    select(1, 361)
  xyz_df <- raw_df %>%
    setNames(paste0('x', 0:360)) %>%
    ### fix discontinuity
    mutate(x0 = (x0 + x360)/2) %>%
    mutate(lat = 90:-90) %>%
    ### drop north pole; cells will shift upward by .5 deg
    filter(lat != 90) %>%
    gather(lon, microplastic_mass, -lat) %>%
    mutate(lon = as.numeric(str_remove(lon, 'x'))) %>%
    select(lon, lat, microplastic_mass) %>%
    ### align cell centers
    mutate(lat = lat + 0.5, 
           lon = lon + 0.5) %>%
    ### rotate 0 to 360 to become -180 to 180
    mutate(lon = ifelse(lon > 180, -360 + lon, lon))
  
  ### plot to check alignment of cells:
  # land_sf <- read_sf(here('_spatial/ne_10m_land/ne_10m_land.shp'))
  # ggplot() +
  #   geom_raster(data = xyz_df,
  #               aes(x = lon, y = lat, fill = microplastic_mass),
  #               na.rm = TRUE) +
  #   geom_sf(data = land_sf, size = .1, alpha = .5) +
  #   scale_fill_viridis_c(na.value = NA)
  
  plast_r <- rasterFromXYZ(xyz_df, crs = '+init=epsg:4326')
  
  ### run one level of focal on this to fill potential gaps
  focal_mtx <- matrix(1, 3, 3)
  plas_gf_r <- focal(plast_r, focal_mtx, 
                     fun = mean, na.rm = TRUE, 
                     NAonly = TRUE, pad = TRUE,
                     progress = 'text')
  ### plot to check alignment of cells:
  # land_sf <- read_sf(here('_spatial/ne_10m_land/ne_10m_land.shp'))
  # ggplot() +
  #   geom_raster(data = as.data.frame(plas_gf_r, xy = TRUE) %>% filter(!is.na(layer)),
  #               aes(x = x, y = y), fill = 'red',
  #               na.rm = TRUE) +
  #   geom_raster(data = as.data.frame(plast_r, xy = TRUE),
  #               aes(x = x, y = y, fill = microplastic_mass),
  #               na.rm = TRUE) +
  #   geom_sf(data = land_sf, size = .1, alpha = .5) +
  #   scale_fill_viridis_c(na.value = NA)

  
  ### Reproject to Mollweide 10 km x 10 km
  plas_gf_mol_r <- plas_gf_r %>%
    projectRaster(ocean_area_mol) %>%
    mask(ocean_area_mol)
    
  plas_gf_mol_resc_r <- rescale_stressor(plas_gf_mol_r, qtile = .999)

  ### round to drop ludicrous precision then save out
  values(plas_gf_mol_resc_r) <- round(values(plas_gf_mol_resc_r), 5)
  writeRaster(plas_gf_mol_resc_r, microplastics_f, overwrite = TRUE)
  
}

r <- raster(microplastics_f)
  
plot(r, axes = FALSE, 
     main = 'rescaled microplastic mass')

hist(r[r > 0], main = 'distribution of non-zero values')

```


## Climate stressors other than long-term temperature rise

### Sea surface temperature extremes

Analogous to marine heat waves.  

* dir: `stressors_2021/_dataprep/SST_past_present/final_extreme`
        * file: `sst_exceeds_90quant_2016-2020.tif`
        
Read in, leave as proportion of weeks exceeding 90th percentile of climatological variance over the reference period (no need for rescale).

```{r}
x_sst_local_f <- here_anx('stressors/max_temp', 
                          'sst_exceeds_90quant_2016-2020.tif')
if(!file.exists(x_sst_local_f)) {
  extreme_sst_orig_f <- here_ohi('stressors_2021/_dataprep/SST_past_present/final_extreme', 
                            'sst_exceeds_90quant_2016-2020.tif')
  file.copy(extreme_sst_orig_f, x_sst_local_f)
}

extreme_sst_r <- raster(x_sst_local_f)

compareRaster(extreme_sst_r, ocean_rast)

### round values to reduce file size
values(extreme_sst_r) <- round(values(extreme_sst_r), 3)

writeRaster(extreme_sst_r, here('_data/stressors_mol/sst_extremes_2020.tif'), overwrite = TRUE)
```

### Sea level rise from OHI 2021

* Use annual mean SLR anomaly rasters from here:
    * dir: `git-annex/globalprep/prs_slr/v2021/int/msla_annual_mean/`
        * files: `msla_annual_2015.tif` to `msla_annual_2019.tif`
    * These are pre-transform and mask.  Transformed and masked are also available, but let's use these so we can make sure the transform and mask are accurate to our particular CRS.
* See OHI processing script for details: `https://github.com/OHI-Science/ohiprep_v2021/blob/gh-pages/globalprep/prs_slr/v2021/slr_layer_prep_v2.Rmd`

```{r Sea level rise data}
slr_f <- here('_data/stressors_mol/sea_level_rise_2019.tif')
### unlink(slr_f)

if(!file.exists(slr_f)) {
  years <- 2015:2019
  local_fs <- here_anx('stressors/slr', sprintf('msla_annual_%s.tif', years))
  if(any(!file.exists(local_fs))) {
    orig_fs <- here_ohi('git-annex/globalprep/prs_slr/v2021/int/msla_annual_mean/', 
                        sprintf('msla_annual_%s.tif', years))
    file.copy(orig_fs, local_fs)
  }
  raw_stack <- raster::stack(local_fs)
  
  raw_r <- calc(raw_stack, mean)
  # class      : RasterLayer 
  # dimensions : 720, 1440, 1036800  (nrow, ncol, ncell)
  # resolution : 0.25, 0.25  (x, y)
  # extent     : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # source     : msla_annual_2019.tif 
  # names      : msla_annual_2019 
  # values     : -1.201033, 0.6790083  (min, max)
  
  ### In WGS84, .25 degree, so large enough res for projectRaster directly.
  ### First, use focal interpolation of mean of neighboring cells within 
  ### a degree (inverse distance weighting) to help make raw SLR layer 
  ### extends all the way to coastline.
  focal_rad <- 4
  focal_dia <- focal_rad*2+1
  focal_mtx <- matrix(nrow = focal_dia, ncol = focal_dia)
  for(i in 1:focal_dia) {
    for(j in 1:focal_dia) {
      ### i <- 1; j <- 1
      d <- sqrt((i - (focal_rad + 1))^2 + (j - (focal_rad + 1))^2)
      focal_mtx[i, j] <- 1 / d
    }
    focal_mtx[focal_rad + 1, focal_rad + 1] <- 1
  }
  gf_r <- focal(raw_r, focal_mtx, 
                fun = mean, na.rm = TRUE, progress = 'text')

  ### replace NA cells in the original with their counterparts in gf_r_only
  combo <- raw_r
  values(combo)[is.na(values(combo))] <- values(gf_r)[is.na(values(combo))]
  
  ### Project to Mollweide 10km and masking to coastal 1 km cells
  slr_moll_r <- projectRaster(combo, ocean_rast, progress = 'text')
  coastal_r  <- raster(here('_spatial/coastal_1km_area_mol.tif'))

  # slr_coastal_r <- mask(slr_moll_r, coastal_r, progress = 'text')
  slr_coastal_r <- slr_moll_r * coastal_r
  # missing <- coastal_r
  # values(missing)[!is.na(values(slr_coastal_r))] <- NA
  # plot(missing, col = 'red')
  # ### still 4161 coastal cells missing values - polar regions
  
  ### Set negative values to zero, then rescale
  values(slr_coastal_r)[values(slr_coastal_r) < 0] <- 0
  slr_coastal_r <- rescale_stressor(slr_coastal_r, qtile = .999)

  ### round to drop ludicrous precision then save out
  values(slr_coastal_r) <- round(values(slr_coastal_r), 5)
  writeRaster(slr_coastal_r, slr_f, overwrite = TRUE)
  
  ### to identify post-mask gapfilled cells
  # gf_map_wgs84 <- mask(gf_r, orig_r, inverse = TRUE)
  # gf_map_moll <- projectRaster(gf_map_wgs84, ocean_rast)
  # gf_map_moll_mask <- mask(gf_map_moll, coastal_r, inverse = FALSE)
  # plot(gf_map_moll_mask, col = 'red')
  
}

r <- raster(slr_f)
  
plot(r, axes = FALSE,
     main = 'Sea level rise rescaled to 99.9%ile')

hist(r[r > 0], main = 'distribution of non-zero values')

```
    
### Ocean acidification from CHI Pace of Change (2017 data year)

* dir: `git-annex/globalprep/prs_oa/v2017/int/annual_avg_moll_rescaled_int`
    * file: `oa_resc_int_2020.tif`
* These are in Mollweide and rescaled based on normalized gain toward $\Omega = 1$: 
    * $stressor_{OA} = \begin{cases} \frac{\Omega_{ref} - \Omega_{2020}}{\Omega_{ref} - 1} \text{ when } \Omega_{2020} > 1\\ 1 \text{ when } \Omega_{2020} \leq 1 \end{cases}$
    * The resolution is based on the original resolution; the data have been gapfilled.
    * Here, resample to 10 km Mollweide and mask out land

```{r Ocean acidification data}
oa_f <- here('_data/stressors_mol/ocean_acidification_2020.tif')
### unlink(oa_f)

if(!file.exists(oa_f)) {
  
  oa_resc_mol_local_f <- here_anx('stressors/oa', 'oa_resc_int_2020.tif')
  if(!file.exists(oa_resc_mol_local_f)) {
    oa_resc_mol_f <- here_ohi('git-annex/globalprep/prs_oa/v2017/int', 
                         'annual_avg_moll_rescaled_int',
                         'oa_resc_int_2020.tif')
    file.copy(oa_resc_mol_f, oa_resc_mol_local_f)
  }
  raw_r <- raster(oa_resc_mol_f)
  # class      : RasterLayer 
  # dimensions : 341, 332, 113212  (nrow, ncol, ncell)
  # resolution : 112000, 54700  (x, y)
  # extent     : -18597502, 18586498, -9085652, 9567048  (xmin, xmax, ymin, ymax)
  # crs        : +proj=moll +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +units=m +no_defs 
  # source     : oa_resc_int_2020.tif 
  # names      : oa_resc_int_2020 
  # values     : 0, 1  (min, max)
  
  ### project to 10 km
  oa_r <- projectRaster(raw_r, ocean_rast, progress = 'text')
  oa_masked <- mask(oa_r, ocean_rast)
  
  ### round to drop ludicrous precision then save out
  values(oa_masked) <- round(values(oa_masked), 4)
  writeRaster(oa_masked, oa_f, overwrite = TRUE)
  
}

r <- raster(oa_f)
  
plot(r, axes = FALSE,
     main = 'Ocean acidification rescaled by normalized loss toward Ω = 1')

hist(r[r > 0], main = 'distribution of non-zero values')

```
    
    
### Ultraviolet radiation from OHI 2021

* dir: `git-annex/globalprep/prs_uv/v2021/int/rescaled/`
    * file: `annual_diff_2016_2020_rescaled.tif`

```{r Ultraviolet radiation data}
uv_f <- here('_data/stressors_mol/uv_radiation_2020.tif')
### unlink(uv_f)

if(!file.exists(uv_f)) {
  local_f <- here_anx('stressors/uv', 'annual_diff_2016_2020_rescaled.tif')
  if(!file.exists(local_f)) {
    orig_f <- here_ohi('git-annex/globalprep/prs_uv/v2021/int/rescaled', 
                       'annual_diff_2016_2020_rescaled.tif')
    file.copy(orig_f, local_f)
  }
  raw_r <- raster(local_f)
  # class      : RasterLayer 
  # dimensions : 180, 360, 64800  (nrow, ncol, ncell)
  # resolution : 1, 1  (x, y)
  # extent     : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # source     : annual_diff_2016_2020_rescaled.tif 
  # names      : annual_diff_2016_2020_rescaled 
  # values     : 0, 1  (min, max)
  
  ### No need to aggregate - native res much greater than analysis res
  uv_r <- projectRaster(raw_r, ocean_rast, progress = 'text')
  uv_r <- mask(uv_r, ocean_rast)
  
  ### round to drop ludicrous precision then save out
  values(uv_r) <- round(values(uv_r), 5)
  writeRaster(uv_r, uv_f, overwrite = TRUE)
}

r <- raster(uv_f)
  
plot(r, axes = FALSE,
     main = 'Ultraviolet radiation anomalies')

hist(r[r > 0], main = 'distribution of non-zero values')
```

## Ocean-based stressors


## Sensory stressors

### Light pollution

Per the suggestion by the authors, we will clip out any values below a threshold (7? 20?) as being prone to fluctuation and therefore potentially not stable.  

> The VIIRS-derived results are more fluctuated during the period from 2015 to 2018 for pixels with DN values greater than 7 (Fig. 5a,b), and this fluctuation was notably mitigated for pixels with relatively high DNs. This suggests the converted DN values from VIIRS data are more reliable for pixels with DN values larger than 20 (Fig. 5c,d), and the derived temporal patterns of SNTL and lit pixels are more reliable for pixels with DN values above 30 (Fig. 5e,f). There are two possible reasons for the fluctuation of derived NTL time series data from VIIRS. First, the original observations of VIIRS NTL data fluctuate over years24. For example, the drop of SNTL and lit pixels from 2015 to 2016 in our result was also observed in the annual result of VIIRS data from the Payne Institute for Public Policy under the Colorado School of Mines24. Second, the simulated DMSP-like NTL data from VIIRS have a larger extent than the DMSP data with the improved sensitivity of sensors in VIIRS and the aggregation procedure using the kernel density method and point-spread function, especially in low luminance regions (Fig. S3). After excluding those low luminance regions using the threshold of 7, the derived NTL result from VIIRS is closer with the DMSP data. When the threshold increases to 20 and 30, most blooming effects around the city and villages are eliminated, resulting in a more continuous time series data of the SNTL and the number of lit pixels (Fig. 5)

Note that with a threshold of 7 there appear to be artifacts in the 2018 data in the arctic and in the waters south of Australia.  Could this be legit light pollution, or auroras?  Try using the minimum value for each cell between the 2017 and 2018 rasters, then dropping values below 10.

Also, examine the ocean-based sources of stable light as potential benthic structures?

* dir: `/home/shares/ohi/spp_vuln/spp_vuln_mapping/stressors/light`
    * file: `Harmonized_DN_NTL_2018_simVIIRS.tif`
    
```{r light pollution}
light_f <- here('_data/stressors_mol/light_2018.tif')
### unlink(light_f)

if(!file.exists(light_f)) {
  light_fs <- list.files(here_anx('stressors/light'),
                         full.names = TRUE)
  light_stack <- stack(light_fs)
  # class      : RasterLayer 
  # dimensions : 16801, 43201, 725820001  (nrow, ncol, ncell)
  # resolution : 0.008333333, 0.008333333  (x, y)
  # extent     : -180.0042, 180.0042, -65.00417, 75.00417  (xmin, xmax, ymin, ymax)
  # crs        : +proj=longlat +datum=WGS84 +no_defs 
  # source     : Harmonized_DN_NTL_2018_simVIIRS.tif 
  # names      : Harmonized_DN_NTL_2018_simVIIRS 
  # values     : 0, 63  (min, max)
  
  ### take min of 2017 and 2018 to help drop artifacts
  min_light_r <- calc(light_stack, fun = min, progress = 'text')
  
  ### crop north/south slightly to remove artifacts
  bbox <- c(-180, 180, -64.5, 74.5)
  crop_light_r <- crop(min_light_r, extent(bbox), progress = 'text')

  thresh <- 10
  values(crop_light_r)[values(crop_light_r) < thresh] <- 0
  
  light_agg_r <- aggregate(crop_light_r, fact = 12, progress = 'text')
  
  light_r_mol <- projectRaster(light_agg_r, ocean_rast, progress = 'text')
  ### fill NA cells with zero
  values(light_r_mol)[is.na(values(light_r_mol))] <- 0
  ### convert negative cells (result of projection artifacts?) to zero
  values(light_r_mol)[values(light_r_mol) < 0] <- 0
  light_r_mol <- mask(light_r_mol, ocean_rast) %>%
    rescale_stressor(qtile = .999)
  
  ### round to drop ludicrous precision then save out
  values(light_r_mol) <- round(values(light_r_mol), 5)
  writeRaster(light_r_mol, light_f, overwrite = TRUE)
}

r <- raster(light_f)
  
plot(r, axes = FALSE,
     main = 'Light pollution (nighttime lights)')

hist(r[r > 0], main = 'distribution of non-zero values')
```
